sandbox fixes and updates to it

+++ app/sandbox/docker_sandbox.py
@@ -8,7 +8,9 @@ from pathlib import Path
 
 from loguru import logger
 
+logger.remove()
 logger.add("app\\logs\\run.log") # log file where logs will be added to
+
 SANDBOX_DIR = Path(__file__).parent.resolve() # explicit path to locate Dockerfile
 
 def set_sandbox_img():
@@ -17,12 +19,17 @@ def set_sandbox_img():
     try:
         image = client.images.get("agent-sandbox-container:early")
         print("image container exist!")
+        logger.info("image container exist!")
         
         return True
         
     except docker.errors.ImageNotFound:
         print("image not found, building the image container....")
+        logger.info("image not found, building the image container....")
+        
         print("path = ",str(SANDBOX_DIR))
+        logger.info("path = ",str(SANDBOX_DIR))
+        
         image,log = client.images.build(
                                         path=str(SANDBOX_DIR),
                                         dockerfile="Dockerfile",
@@ -30,6 +37,8 @@ def set_sandbox_img():
                                         rm=True,)
         
         print(f"image built, image id = {image.id}")
+        logger.info(f"image built, image id = {image.id}")
+        
         return True
 
 def gen_temp_file(agent_code:str) -> str:
@@ -38,15 +47,24 @@ def gen_temp_file(agent_code:str) -> str:
     
     if not code_match:
         print(">>> No Python code found in the response.")
+        logger.info(">>> No Python code found in the response.")
+        
         return False
 
     code_extracted = code_match.group(1)
     
-    with tempfile.NamedTemporaryFile(mode="w",suffix=".py",delete=False) as tmpFile:
+    # isolated workspace folder for sandbox!
+    sandbox_workspace = os.path.join(tempfile.gettempdir(),"sandbox_workspace")
+    os.makedirs(sandbox_workspace,exist_ok=True)
+    
+    with tempfile.NamedTemporaryFile(mode="w",suffix=".py",
+                                     delete=False, dir=sandbox_workspace) as tmpFile:
+        
         tmpFile.write(code_extracted)
         temp_file_path = tmpFile.name
         
         print("temp file path is = ",temp_file_path)
+        logger.info(f"temp file path is = {temp_file_path}")
 
     return temp_file_path
 
@@ -60,11 +78,16 @@ def execute_code(temp_file_path:str):
                             command=f"python /app/{Path(temp_file_path).name}",
                             volumes={os.path.dirname(temp_file_path):{"bind": "/app","mode":"rw"}},
                             working_dir="/app",
+                            #mem_limit="512m",       # Cap RAM usage to 512MB
+                            #nano_cpus=1000000000,    # Cap CPU usage to 1 core
+                            network_mode="none", # disables network access
                             remove=True)
         
         return container.decode("utf-8")
     except Exception as e:
+        logger.info(f"failed running the container : {e}")
         return f"failed running the container : {e}"
     finally:
-        os.unlink(temp_file_path) #  removes the pointer, used to delete the temp python file
+        if os.path.exists(temp_file_path):
+            os.unlink(temp_file_path) #  removes the pointer, used to delete the temp python file
